Merge the six gated tightening commits through 622d2521. Preserve the
completed master campaign, regenerate compatibility evidence from the
combined source tree, and retain both histories in a first-class merge.

+++ internal/inference/active_qwen_recipe_integration_test.go
@@ -42,7 +42,7 @@ func TestActiveRecipeQwen35Open(t *testing.T) {
 	}
 	dense, recurrent := false, false
 	for index := range runner.program.Model.LayerCount() {
-		layer, layerErr := runner.program.Model.Layer(index)
+		layer, layerErr := runner.program.Model.Layer(uint32(index))
 		if layerErr != nil {
 			t.Fatal(layerErr)
 		}
